more environments, newer regex, automatic wan setup, updated makefiles
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@5312 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/basics/basebuffer.h
@@ -1611,7 +1611,7 @@ public:
 			this->cRpp+=4;
 			return  ip;
 		}
-		return INADDR_ANY;
+		return (uint32)INADDR_ANY;
 	}
 	///////////////////////////////////////////////////////////////////////////
 	sint64 operator = (const sint64 lx)
